+++ pcs/cli/cluster/command.py
@@ -9,6 +9,10 @@ from pcs.cli.resource.parse_args import(
     parse_create_simple as parse_resource_create_args
 )
 from pcs.cli.common.errors import CmdLineInputError
+from pcs.cli.common.parse_args import prepare_options
+from pcs.cli.common.console_report import error
+from pcs.lib.errors import LibraryError
+from pcs.common import report_codes
 
 def _node_add_remote_separate_host_and_name(arg_list):
     node_host = arg_list[0]
@@ -44,3 +48,34 @@ def node_add_remote(lib, arg_list, modifiers):
         allow_invalid_instance_attributes=force,
         wait=modifiers["wait"],
     )
+
+def node_add_guest(lib, arg_list, modifiers):
+    if len(arg_list) < 2:
+        raise CmdLineInputError()
+
+    meta_options = prepare_options(arg_list[2:])
+    if "remote-node" in meta_options:
+        error("option 'remote-node' is not allowed")
+        raise CmdLineInputError()
+
+    meta_options.update({"remote-node": arg_list[0]})
+    resource_id = arg_list[1]
+
+    force = modifiers["force"]
+
+    try:
+        lib.cluster.node_add_guest(
+            resource_id,
+            meta_options,
+            allow_incomplete_distribution=force,
+            allow_pacemaker_remote_service_fail=force,
+            wait=modifiers["wait"],
+        )
+    except LibraryError as e:
+        for report in e.args:
+            if report.code == report_codes.INVALID_OPTION:
+                report.info["allowed"] = [
+                    value for value in report.info["allowed"]
+                    if value != "remote-node"
+                ]
+        raise e