+++ pcs/lib/cib/resource/__init__.py
@@ -10,6 +10,7 @@ from pcs.lib.cib.resource import (
     clone,
     common,
     group,
+    guest_node,
     operations,
     primitive,
 )