+++ pcs/lib/cib/resource/common.py
@@ -5,6 +5,8 @@ from __future__ import (
     unicode_literals,
 )
 
+from functools import partial
+
 from pcs.lib.cib import nvpair
 from pcs.lib.cib.resource.bundle import (
     is_bundle,
@@ -200,3 +202,19 @@ def unmanage(resource_el):
             "is-managed": "false",
         }
     )
+
+arrange_first_meta_attributes = partial(
+    nvpair.arrange_first_nvset,
+    "meta_attributes"
+)
+
+def has_meta_attribute(resource_el, name):
+    """
+    Return if the element contains meta attribute 'name'
+
+    etree.Element resource_el is researched element
+    string name specifies attribute
+    """
+    return 0 < len(resource_el.xpath(
+        './meta_attributes/nvpair[@name="{0}"]'.format(name)
+    ))